requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
blocklist
flask-cors
numpy
pandas
scikit-learn
joblib
pytest
